Sanctuaryside admin CRUD

+++ StainedGlass.Web/src/admin/app.ts
@@ -1,21 +1,26 @@
 import 'es6-promise/auto'; 
 
 //handle church deleting
-document.querySelectorAll('.deleteChurch').forEach( button => {
+document.querySelectorAll('.delete').forEach( button => {
+    
     var slug = button.getAttribute('data-slug');
+    var type = button.getAttribute('data-type');
     // @ts-ignore
     button.addEventListener('click', async (): Promise<void> => {
-       const response = await fetch('/admin/church/' + slug, {
-           method: 'DELETE',
-           headers: { 'Content-Type': 'application/json' }
-       });
-        if (response.ok) 
-        { 
-            location.reload();
-            } 
-        else 
-        { 
-            console.error('Delete request failed'); 
+        let userConfirmed = confirm("Are you sure you want to delete this item?");
+        if (userConfirmed) {
+            const response = await fetch('/admin/' + type + '/' + slug, {
+                method: 'DELETE',
+                headers: { 'Content-Type': 'application/json' }
+            });
+            if (response.ok)
+            {
+                location.reload();
+            }
+            else
+            {
+                console.error('Delete request failed');
+            }
         }
     });
 });
@@ -23,42 +28,42 @@ document.querySelectorAll('.deleteChurch').forEach( button => {
 document.addEventListener('DOMContentLoaded', () => {
     //handle church editing
     const editForm = document.getElementById('editForm') as HTMLFormElement;
-    const churchSlug = editForm.getAttribute('data-slug');
-    const redirectUrl = editForm.getAttribute('data-redirect-url') as string;
-    editForm.addEventListener('submit', async (event): Promise<void> => {
-        event.preventDefault();
-        const formData = new FormData(editForm);
-        // const data: { [key: string]: string } = {};
-        // formData.forEach((value, key) => {
-        //     data[key] = value as string;
-        // });
-        // //if an image is uploaded - convert it to base64
-        // const file = formData.get('Image') as File;
-        // if (file.name.length > 0)
-        // {
-        //     const base64File = await convertFileToBase64(file);
-        //     data.Image = base64File;
-        // } else
-        // {
-        //     //if not - do not send it to the server;
-        //     delete data.Image;
-        // }
-        
-        // console.log(data);
+    if (editForm != null)
+    {
+        const churchSlug = editForm.getAttribute('data-slug');
+        const redirectUrl = editForm.getAttribute('data-redirect-url') as string;
+        editForm.addEventListener('submit', async (event): Promise<void> => {
+            event.preventDefault();
+            const formData = new FormData(editForm);
 
-        const response = await fetch('/church/' + churchSlug, { 
-            method: 'PUT',
-            body: formData 
-        }); 
-        if (response.ok) 
-        { 
-            location.href = redirectUrl;
-        } 
-        else 
-        { 
-            console.error('Error submitting edit form'); 
-        }
-    });
+            const response = await fetch('/church/' + churchSlug, {
+                method: 'PUT',
+                body: formData
+            });
+            if (response.ok)
+            {
+                location.href = redirectUrl;
+            }
+            else
+            {
+                console.error('Error submitting edit form');
+            }
+        });
+    }
+    
+    //select an item in the church select (sanctuarySide page)
+    const addForm = document.getElementById('addForm') as HTMLFormElement;
+    if (addForm != null)
+    {
+        addForm.addEventListener('submit', (event)  => {
+            const churchSlug = document.getElementById('churchSlug') as HTMLSelectElement;
+            if (churchSlug.value == "0")
+            {
+                event.preventDefault();
+                alert("You must select an item in the Church select");
+            }
+        });
+    }
 });
 
 async function convertFileToBase64(file: File): Promise<string>
